Arreglo de vista DatosClient(separacion en componetes
