+++ src/components/stadisticalData/CardDataClient/cardDataClient.module.css
@@ -1,0 +1,5 @@
+.cardList {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-around;
+}
