+++ src/components/stadisticalData/stadisticalData.module.css
