+++ src/container/App.jsx
@@ -13,7 +13,8 @@ import Admin from '../pages/Admin/admin';
 import AddAdmin from '../pages/AddAdmin/addAdmin';
 import AddDevice from '../pages/AddDevice/addDevice';
 import SpecificGraph from '../pages/SpecificGraph/specificGraph';
-import Data from '../pages/stadisticalData/stadisticalData';
+import DataClient from '../pages/StadisticalDataClient/dataClient';
+import Data from '../components/stadisticalData/stadisticalData';
 import { PlantProvider } from '../components/PlantContext/plantContext';
 function App() {
   const [isAuthenticated, setIsAuthenticated] = useState(false);
@@ -40,9 +41,10 @@ function App() {
           <Route path='/specific/:plantName' element={<SpecificPlant />} />
           <Route path='/graph' element={<Graph />} />
           <Route path='/adminData' element={<AdminData />} />
-          <Route path="/data" element={<Data />} />
+          <Route path="/dataClient" element={<DataClient />} />
           <Route path='/addAdmin' element={<AddAdmin />} />
           <Route path='/specificGraph/' element={<SpecificGraph />} />
+          <Route path='/stadistical/:plantName' element={<Data />} />
           <Route path="/agregarplanta" element={<AgregarPlanta />} />
           <Route path="/agregarplantaadmin" element={<AgregarPlantaAdmin />} />
           <Route path="/addDevice" element={<AddDevice />} />
